+++ test/foundry/unit/UniswapV3Factory/setFeeManager.t.sol
@@ -1,7 +1,7 @@
 pragma solidity ^0.7.6;
 pragma abicoder v2;
 
-import {UniswapV3FactoryTest} from './UniswapV3Factory.t.sol';
+import {UniswapV3FactoryTest} from "./UniswapV3Factory.t.sol";
 
 contract SetFeeManagerTest is UniswapV3FactoryTest {
     function setUp() public override {