+++ test/foundry/unit/UniswapV3Pool/init.t.sol
@@ -1,8 +1,8 @@
 pragma solidity ^0.7.6;
 pragma abicoder v2;
 
-import {IUniswapV3Pool} from 'contracts/core/interfaces/IUniswapV3Pool.sol';
-import {UniswapV3PoolTest} from './UniswapV3Pool.t.sol';
+import {IUniswapV3Pool} from "contracts/core/interfaces/IUniswapV3Pool.sol";
+import {UniswapV3PoolTest} from "./UniswapV3Pool.t.sol";
 
 contract InitTest is UniswapV3PoolTest {
     function test_RevertIf_AlreadyInit() public {